chore: parser helpers next to their users
