--- packages/svelte/src/compiler/phases/1-parse/utils/create.js
@@ -1,16 +1,0 @@
-/** @import { AST } from '#compiler' */
-
-/**
- * @param {any} transparent
- * @returns {AST.Fragment}
- */
-export function create_fragment(transparent = false) {
-	return {
-		type: 'Fragment',
-		nodes: [],
-		metadata: {
-			transparent,
-			dynamic: false
-		}
-	};
-}
