+++ packages/svelte/src/compiler/phases/2-analyze/visitors/SvelteBody.js
@@ -2,7 +2,7 @@
 /** @import { Context } from '../types' */
 import * as e from '../../../errors.js';
 import { is_event_attribute } from '../../../utils/ast.js';
-import { disallow_children } from './shared/special-element.js';
+import { disallow_children } from '../../nodes.js';
 import { check_global_event_reference } from './shared/utils.js';
 
 /**
