+++ packages/svelte/src/compiler/phases/2-analyze/visitors/SvelteWindow.js
@@ -1,6 +1,6 @@
 /** @import { AST } from '#compiler' */
 /** @import { Context } from '../types' */
-import { disallow_children } from './shared/special-element.js';
+import { disallow_children } from '../../nodes.js';
 import * as e from '../../../errors.js';
 import { is_event_attribute } from '../../../utils/ast.js';
 import { check_global_event_reference } from './shared/utils.js';
