+++ packages/svelte/src/compiler/phases/nodes.js
@@ -1,6 +1,7 @@
 /** @import { Expression, PrivateIdentifier, SourceLocation } from 'estree' */
 /** @import { AST, Binding } from '#compiler' */
 import * as b from '#compiler/builders';
+import * as e from '../errors.js';
 
 /**
  * All nodes that can appear elsewhere than the top level, have attributes and can contain children
@@ -43,6 +44,35 @@ export function is_custom_element_node(node) {
 		(node.name.includes('-') ||
 			node.attributes.some((attr) => attr.type === 'Attribute' && attr.name === 'is'))
 	);
+}
+
+/**
+ * @param {boolean} transparent
+ * @returns {AST.Fragment}
+ */
+export function create_fragment(transparent = false) {
+	return {
+		type: 'Fragment',
+		nodes: [],
+		metadata: {
+			transparent,
+			dynamic: false
+		}
+	};
+}
+
+/**
+ * @param {AST.SvelteBody | AST.SvelteDocument | AST.SvelteOptionsRaw | AST.SvelteWindow} node
+ */
+export function disallow_children(node) {
+	const { nodes } = node.fragment;
+
+	if (nodes.length > 0) {
+		const first = nodes[0];
+		const last = nodes[nodes.length - 1];
+
+		e.svelte_meta_invalid_content({ start: first.start, end: last.end }, node.name);
+	}
 }
 
 /**
