+++ packages/svelte/src/compiler/utils/fuzzymatch.js
