+++ packages/svelte/src/compiler/utils/mapped_code.js
@@ -1,16 +1,28 @@
 /** @import { ValidatedCompileOptions } from '#compiler' */
+
 /** @import { Processed } from '../preprocess/public.js' */
 /** @import { SourceMap } from 'magic-string' */
 /** @import { Source } from '../preprocess/private.js' */
 /** @import { DecodedSourceMap, SourceMapSegment, RawSourceMap } from '@jridgewell/remapping' */
 import remapping from '@jridgewell/remapping';
-import { push_array } from './push_array.js';
 
 /**
  * @param {string} s
  */
 function last_line_length(s) {
 	return s.length - s.lastIndexOf('\n') - 1;
+}
+
+/**
+ * `array.push(...items)` overflows the stack for large `items`.
+ * @template T
+ * @param {T[]} array
+ * @param {T[]} items
+ */
+function push_array(array, items) {
+	for (let i = 0; i < items.length; i++) {
+		array.push(items[i]);
+	}
 }
 // mutate map in-place
